requirements.txt
fastapi
uvicorn
pydantic
llama-cpp-python==0.2.23
